Setting semantics as a property of STL_Formula class
Instead of using a global variable "objToUse" as before, we now have
a property of the STL_Formula class called "semantics".

To set the semantics of an STL formula, use the set_semantics function.
Similarly, one can get the semantics of an STL formula by using get_semantics.

All the other updated functions now use the new convention of semantics as a
property of the STL formula class, rather than the old way of using a global variable.