No longer removing identical points in RobustAlways
The reason for not removing identical points in RobustAlways any longer is
becuase the additive semantics is dependent on the sampling interval (i.e.
changing the time between sampled points can change the robustness). In an
attempt to make the robustness as similar as possible between simulation runs,
we do not remove samples which have identical value to the sample before.

In addition to this, there was a slight indexing issue when looking at the
final index of the current timing window. This has now been fixed.