Adding startFunctionValues as input to falsificationProblem
Now we can send a set of starting function values evaluated on a set of
starting parameter points. For now, this is only implemented for the 'snobfit'
solver in BreachProblem.solve.

Also fixed so that the frequency at which the 'snobfit' solver shows updates is
determined by the 'freq_update' parameter of the BreachProblem class.