Use batches for parallel calculation in BreachProblem
In an effort to not run out of memory for parallel calculations,
we introduce a "batch size" for parallel calculations (in BreachProblem.solve).

The idea is that the solver will run a number of simulations in parallel
equal to the batch size, then store the results, then keep running the next
batch, etc. This continues until all requested simulations have been run.