Fixed small bug with time sensitivity in RobustAlways.m
A bug showed up when trying to get startIdxNew and it turned out that
startTime and timeValues at a specific place had very very close values,
but not equal.

During the check "if time_values(endCounter) > endTime", it so happened
that time_values(endCounter) was 27.16, and endTime was around 27.16000000001.

To fix this issue, we now check whether they are "close enough" (which is
arbitrarily chosen as "within 1e-10 of each other"). This fixed the bug for
the given example, which was that phi_RFC_act ALWAYS had robustness 0
when simulating the benchmark AT_and_specifications.

In addition to this change, we also now remove an entry in
time_values and valarray if BOTH of the following hold:
 valarray(entry_index) == valarray(entry_index -1 )
 time_values(entry_index) == time_values(entry_index - 1)